add hooks for improved state handling
